% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export_CSVMM.R
\name{Export_CSVMM}
\alias{Export_CSVMM}
\title{Export Merged Models - CSV}
\usage{
Export_CSVMM()
}
\value{
CSV files with the models
}
\description{
The Exports_CSVMM function exports all merged models created.
}
