% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export_XLSXMM.R
\name{Export_XLSXMM}
\alias{Export_XLSXMM}
\title{Export Merged Models - XLSX}
\usage{
Export_XLSXMM()
}
\value{
Excel files with the models
}
\description{
The Export_XLSXMM function exports all merged models created.
}
