% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_xlsx_files.R
\name{import_xlsx_files}
\alias{import_xlsx_files}
\title{Import all XLSX files from a path}
\usage{
import_xlsx_files(path)
}
\value{
A list of data frames
}
\description{
Import all XLSX files from a path
}
