% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_Models.R
\name{Merge_Models}
\alias{Merge_Models}
\title{Merge Models create by Run_RothCO2}
\usage{
Merge_Models(Models = "All", Export = "Both")
}
\arguments{
\item{Models}{From the various models (folders) created in the function Run_RothCO2, choose the ones to merge. Defaults to "All", but can also be: "VXC", "VXP", "YC", "YCO2D", "YCO2ND", "YCT". "VXC" considers a continuous input of C each year (1tC/year). "VXP": considers a one-off input of C in the first year (1tC/fist-year). "YC": yearly C emissions considering a GWP100 with delay (considering a one-off input of 1tC in the first year). "YCO2D": yearly CO2 emissions considering a GWP100 with delay (considering a one-off input of 1tC in the first year). "YCO2ND": yearly CO2 emissions (considering a one-off input of 1tC in the first year). "YCT": yearly C tails (considering a one-off input of 1tC in the first year) - i.e. what remains on the soil.}

\item{Export}{Export format (.csv or .xlsx). Defaults to "Both", but can also be "XLSX" (.xlsx; microsoft excel file format) or "CSV" (.csv; comma separated values format)}
}
\value{
merged models exported in respective folders and merged models saved in the global environment.
}
\description{
The Merge_Models function merge all chosen models by "Models" parameter created by the functions Run_RothC_[].
}
