% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_Models_VXC.R
\name{Merge_VXC}
\alias{Merge_VXC}
\title{Merge VXC Models}
\usage{
Merge_VXC()
}
\value{
A dataframe with all VXC models merged together
}
\description{
The Merge_VXC function merge all VXC models created by the functions Run_RothC_[].
}
