% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_Models_VXP.R
\name{Merge_VXP}
\alias{Merge_VXP}
\title{Merge VXP Models}
\usage{
Merge_VXP()
}
\value{
A mvector with monthly precipitation values
}
\description{
The Merge_VXP function merge all VXP models created by the functions Run_RothC_[].
}
