% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_Models_YC.R
\name{Merge_YC}
\alias{Merge_YC}
\title{Merge YC Models}
\usage{
Merge_YC()
}
\value{
A dataframe with all YC models merged together
}
\description{
The Merge_YC function merge all YC models created by the functions Run_RothC_[].
}
