% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_Models_YCO2D.R
\name{Merge_YCO2D}
\alias{Merge_YCO2D}
\title{Merge YCO2D Models}
\usage{
Merge_YCO2D()
}
\value{
A dataframe with YCO2 Models merged together
}
\description{
The Merge_YCO2D function merges all YCO2D models created by the functions Run_RothC_[].
}
