% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_Models_YCO2ND.R
\name{Merge_YCO2ND}
\alias{Merge_YCO2ND}
\title{Merge YCO2ND Models}
\usage{
Merge_YCO2ND()
}
\value{
A dataframe with all YCO2ND models merged together
}
\description{
The Merge_YCO2D function merge all YCO2ND models created by the functions Run_RothC_[].
}
