% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_Models_YCT.R
\name{Merge_YCT}
\alias{Merge_YCT}
\title{Merge YCT Models}
\usage{
Merge_YCT()
}
\value{
A dataframe with all YCT models merged together
}
\description{
The Merge_YCT function merge all YCT models created by the functions Run_RothC_[].
}
