% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_RothC_All.R
\name{Run_RothC_All}
\alias{Run_RothC_All}
\title{Run the whole Model package}
\usage{
Run_RothC_All(
  Scope,
  SL_years = 100,
  soil.thick = 23,
  Tem = Temp,
  Pre = Precip,
  Eva = Evp,
  CTRM = 1,
  RTRM = 0.93,
  NTRM = 0.95
)
}
\description{
The Set_Temp function create a column with 12 rows and one column. The first
column indicates the Month (1-12). The second column indicates the mean Temperature,
either filled by the user or autofilled by default values for Spain 1990-2020.
Information from here: https://rdrr.io/rforge/SoilR/src/R/fW.RothC.R
}
