% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_RothC_VLC.R
\name{Run_RothC_VLC}
\alias{Run_RothC_VLC}
\title{Run RothC_VLC}
\usage{
Run_RothC_VLC(
  SL_years = 100,
  soil.thick = 23,
  Tem = Temp,
  Pre = Precip,
  Eva = Evp,
  CTRM = 1
)
}
\value{
A dataframe with all VXC models merged together
}
\description{
The Merge_VXC function merge all VXC models created by the functions Run_RothC_[].
}
