% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_RothC_VLR.R
\name{Run_RothC_VLR}
\alias{Run_RothC_VLR}
\title{Run RothC_VLR}
\usage{
Run_RothC_VLR(
  SL_years = 100,
  soil.thick = 23,
  Tem = Temp,
  Pre = Precip,
  Eva = Evp,
  RTRM = 0.93
)
}
\value{
A dataframe with all VXC models merged together
}
\description{
The Merge_VXC function merge all VXC models created by the functions Run_RothC_[].
}
