% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_RothC_VMN.R
\name{Run_RothC_VMN}
\alias{Run_RothC_VMN}
\title{Run RothC_VMN}
\usage{
Run_RothC_VMN(
  SL_years = 100,
  soil.thick = 23,
  Tem = Temp,
  Pre = Precip,
  Eva = Evp,
  NTRM = 0.95
)
}
\value{
A dataframe with all VXC models merged together
}
\description{
The Merge_VXC function merge all VXC models created by the functions Run_RothC_[].
}
