% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_RothCO2.R
\name{Run_RothCO2}
\alias{Run_RothCO2}
\title{Run the RothC Model and quantify C and CO2 fluxes}
\usage{
Run_RothCO2(
  Scope = "All",
  SL_years = 100,
  soil.thick = 23,
  Tem = Temp,
  Pre = Precip,
  Eva = Evp,
  CTRM = 1,
  RTRM = 0.93,
  NTRM = 0.95
)
}
\arguments{
\item{Scope}{Define the model combination that the user is interested in. Defaults to "All", but can also be: "VLC", "VLR", "VLN", "VMC", "VMR", "VMN". L and M stand for Litter and Manure (type of C input). C, R and N stand for Conventional, Reduced and No (Tillage practices).}

\item{SL_years}{Simulation Length of the model in years. Defaults to 100. SL_years doesn't affect the results but the cut-off (when the model stops).}

\item{soil.thick}{Soil thickness in cm. Defaults to 23 as in the original RothC model from Soil R Package.}

\item{Tem}{Temperature dataframe with one column with a sequence 1:12 (months) and the second column with monthly temperature values (in celsius degrees). The function Set_Temp() can be used to create the dataframe, which is then used as a Default.}

\item{Eva}{Evapotrasnpiration dataframe with one column with a sequence 1:12 (months) and the second column with monthly evapotranspiration values (in mm). The function Set_Evp() can be used to create the dataframe, which is then used as a Default.}

\item{CTRM}{Conventional Tillage Rate Modifier. Defaults to 1.00. Source: https://doi.org/10.1016/j.still.2022.105428}

\item{RTRM}{Reduced Tillage Rate Modifier. Defaults to 0.93. Source: https://doi.org/10.1016/j.still.2022.105428}

\item{NTRM}{No Tillage Rate Modifier. Defaults to 0.95. Source: https://doi.org/10.1016/j.still.2022.105428}

\item{Precip}{Precipitation dataframe with one column with a sequence 1:12 (months) and the second column with monthly precipitation values (in mm). The function Set_Precip() can be used to create the dataframe, which is then used as a Default.}
}
\description{
The Run_RothCO2 function is the main function of the Package. It buils up on the RothC-related functions from the package SoilR (Sierra and Mueller 2022) to create six different folder with the following information:
(1) VXC_Models: considers a continuous input of C each year (1tC/year)
(2) VXP_Models: considers a one-off input of C in the first year (1tC/fist-year)
(3) CTailsP: yearly C tails (considering a one-off input of 1tC in the first year) - i.e. what remains on the soil
(4) CO2Emissions_P_NoDelay: yearly CO2 emissions (considering a one-off input of 1tC in the first year)
(5) CO2Emissions_P_Delay: yearly CO2 emissions considering a GWP100 with delay (considering a one-off input of 1tC in the first year)
(6) CEmissions_P: yearly C emissions considering a GWP100 with delay (considering a one-off input of 1tC in the first year)
}
\details{
Besides defining values for some parameters for the RothC Model (Simulation Lenght, Tillage Rate Modifiers, Soil Thickness, etc.), the user can select among 6 different model combinations (parameter "Scope") to calculate the information described above:
(1) Litter, Conventional Tillage, 0-100\% (+10\%) Clay
(2) Litter, Reduced Tillage, 0-100\% (+10\%) Clay
(3) Litter, No Tillage, 0-100\% (+10\%) Clay
(4) Manure, Conventional Tillage, 0-100\% (+10\%) Clay
(5) Manure, Reduced Tillage, 0-100\% (+10\%) Clay
(6) Manure, No Tillage, 0-100\% (+10\%) Clay

If the user set Scope = "All", all following files will be created, sorted in their respective folders (VXC_Models, VXP_Models, CTailsP, CO2Emissions_P_NoDelay, CO2Emissions_P_Delay):

Model Combinations - Litter, Conventional Tillage, 0-100\% (10\%) Clay
VLC0) Litter; 0\%clay; Conventional tillage
VLC1) Litter; 10\%clay; Conventional tillage
VLC2) Litter; 20\%clay; Conventional tillage
VLC3) Litter; 30\%clay; Conventional tillage
VLC4) Litter; 40\%clay; Conventional tillage
VLC5) Litter; 50\%clay; Conventional tillage
VLC6) Litter; 60\%clay; Conventional tillage
VLC7) Litter; 70\%clay; Conventional tillage
VLC8) Litter; 80\%clay; Conventional tillage
VLC9) Litter; 90\%clay; Conventional tillage
VLC10) Litter; 100\%clay; Conventional tillage
Model Combinations - Litter; Reduced Tillage, 0-100\% (+10\%) Clay
VLR0) Litter; 0\%clay; Reduced tillage
VLR1) Litter; 10\%clay; Reduced tillage
VLR2) Litter; 20\%clay; Reduced tillage
VLR3) Litter; 30\%clay; Reduced tillage
VLR4) Litter; 40\%clay; Reduced tillage
VLR5) Litter; 50\%clay; Reduced tillage
VLR6) Litter; 60\%clay; Reduced tillage
VLR7) Litter; 70\%clay; Reduced tillage
VLR8) Litter; 80\%clay; Reduced tillage
VLR9) Litter; 90\%clay; Reduced tillage
VLR10) Litter; 100\%clay; Reduced tillage
Model Combinations - Litter; No Tillage, 0-100\% (+10\%) Clay
VLN0) Litter; 0\%clay; No tillage
VLN1) Litter; 10\%clay; No Tillage
VLN2) Litter; 20\%clay; No Tillage
VLN3) Litter; 30\%clay; No Tillage
VLN4) Litter; 40\%clay; No Tillage
VLN5) Litter; 50\%clay; No Tillage
VLN6) Litter; 60\%clay; No Tillage
VLN7) Litter; 70\%clay; No Tillage
VLN8) Litter; 80\%clay; No Tillage
VLN9) Litter; 90\%clay; No Tillage
VLN10) Litter; 100\%clay; No Tillage
Model Combinations - Manure; Conventional Tillage, 0-100\% (+10\%) Clay
VMC0) Manure; 0\%clay; Conventional tillage
VMC1) Manure; 10\%clay; Conventional tillage
VMC2) Manure; 20\%clay; Conventional tillage
VMC3) Manure; 30\%clay; Conventional tillage
VMC4) Manure; 40\%clay; Conventional tillage
VMC5) Manure; 50\%clay; Conventional tillage
VMC6) Manure; 60\%clay; Conventional tillage
VMC7) Manure; 70\%clay; Conventional tillage
VMC8) Manure; 80\%clay; Conventional tillage
VMC9) Manure; 90\%clay; Conventional tillage
VMC10) Manure; 100\%clay; Conventional tillage
Model Combinations - Manure; Reduced Tillage, 0-100\% (+10\%) Clay
VMR0) Manure; 0\%clay; Reduced tillage
VMR1) Manure; 10\%clay; Reduced tillage
VMR2) Manure; 20\%clay; Reduced tillage
VMR3) Manure; 30\%clay; Reduced tillage
VMR4) Manure; 40\%clay; Reduced tillage
VMR5) Manure; 50\%clay; Reduced tillage
VMR6) Manure; 60\%clay; Reduced tillage
VMR7) Manure; 70\%clay; Reduced tillage
VMR8) Manure; 80\%clay; Reduced tillage
VMR9) Manure; 90\%clay; Reduced tillage
VMR10) Manure; 100\%clay; Reduced tillage
Model Combinations - Manure; No Tillage, 0-100\% (+10\%) Clay
VMN0) Manure; 0\%clay; No tillage
VMN1) Manure; 10\%clay; No Tillage
VMN2) Manure; 20\%clay; No Tillage
VMN3) Manure; 30\%clay; No Tillage
VMN4) Manure; 40\%clay; No Tillage
VMN5) Manure; 50\%clay; No Tillage
VMN6) Manure; 60\%clay; No Tillage
VMN7) Manure; 70\%clay; No Tillage
VMN8) Manure; 80\%clay; No Tillage
VMN9) Manure; 90\%clay; No Tillage
VMN10) Manure; 100\%clay; No Tillage

To merge the models by folder, refer to function Merge_Models() from this package.

Relevant sources:
(1) Falloon, P., Smith, P., Coleman, K., and Marshall, S. (1998). Estimating the size of the inert organic matter pool from total soil organic carbon content for use in the Rothamsted carbon model. Soil Biology and Biochemistry, 30:1207-1211.
(2) Sierra, C.A., Mueller, M., and Trumbore, S.E. (2012). Models of soil organic matter decomposition: the SoilR package, version 1.0. Geosci. Model Dev., 5:1045-1060.
(3) Sierra, C.A., and Mueller, M. (2015). A general mathematical framework for representing soil organic matter dynamics. Ecological Monographs, 85:505-524.
(4) Weihermueller, L., Graf, A., Herbst, M., and Vereecken, H. (2013). Simple pedotransfer functions to initialize reactive carbon pools of the RothC model. European Journal of Soil Science, 64:567-575.
(5) Zimmermann, M., Leifeld, J., Schmidt, M.W.I., Smith, P., and Fuhrer, J. (2007). Measured soil organic matter fractions can be related to pools in the RothC model. European Journal of Soil Science, 58:658-667.
(6) Sierra, C.A. and Mueller, M. (2012). Package 'SoilR', version 1.2.105. Date: 2022-1-21. Documentation available here: https://cran.r-project.org/web/packages/SoilR/SoilR.pdf
(7) Sierra, C.A. (2015). Run the RothC model for a site. Date: 2015-11-19. Available here: https://www.bgc-jena.mpg.de/TEE/basics/2015/11/19/RothC/
(8) Jordon, M. W., & Smith, P. (2022). Modelling soil carbon stocks following reduced tillage intensity: A framework to estimate decomposition rate constant modifiers for RothC-26.3, demonstrated in north-west Europe. Soil and Tillage Research, 222, 105428.
}
