% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Evp.R
\name{Set_Evp}
\alias{Set_Evp}
\title{Set Monthly Evapotranspiration Vector}
\usage{
Set_Evp(
  EvpJan = EvpJan_Default,
  EvpFeb = EvpFeb_Default,
  EvpMar = EvpMar_Default,
  EvpApr = EvpApr_Default,
  EvpMay = EvpMay_Default,
  EvpJun = EvpJun_Default,
  EvpJul = EvpJul_Default,
  EvpAug = EvpAug_Default,
  EvpSep = EvpSep_Default,
  EvpOct = EvpOct_Default,
  EvpNov = EvpNov_Default,
  EvpDec = EvpDec_Default
)
}
\value{
A mvector with monthly evapotranspiration values
}
\description{
The Set_Evp function create a column with 12 rows and one column. The first
column indicates the Month (1-12). The second column indicates the mean monthly evapotranspiration,
either filled by the user or autofilled by default values calculated using the
Thornthwaite equiation using Temperature data from Set_Temp
}
