% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Evp.R
\name{Set_Evp}
\alias{Set_Evp}
\title{Set Monthly Mean Evapotranspirations}
\usage{
Set_Evp(
  EvpJan = EvpJan_Default,
  EvpFeb = EvpFeb_Default,
  EvpMar = EvpMar_Default,
  EvpApr = EvpApr_Default,
  EvpMay = EvpMay_Default,
  EvpJun = EvpJun_Default,
  EvpJul = EvpJul_Default,
  EvpAug = EvpAug_Default,
  EvpSep = EvpSep_Default,
  EvpOct = EvpOct_Default,
  EvpNov = EvpNov_Default,
  EvpDec = EvpDec_Default
)
}
\arguments{
\item{EvpJan}{Mean Potential Evapotranspiration (mm) in January. Default uses values from Temp (can be created by Set_Temp) to calculate Potential Evapotranspiration (mm) using the Thornthwaite equation}

\item{EvpFeb}{Mean Potential Evapotranspiration (mm) in February. Default uses values from Temp (can be created by Set_Temp) to calculate Potential Evapotranspiration (mm) using the Thornthwaite equation}

\item{EvpMar}{Mean Potential Evapotranspiration (mm) in March. Default uses values from Temp (can be created by Set_Temp) to calculate Potential Evapotranspiration (mm) using the Thornthwaite equation}

\item{EvpApr}{Mean Potential Evapotranspiration (mm) in April. Default uses values from Temp (can be created by Set_Temp) to calculate Potential Evapotranspiration (mm) using the Thornthwaite equation}

\item{EvpMay}{Mean Potential Evapotranspiration (mm) in May. Default uses values from Temp (can be created by Set_Temp) to calculate Potential Evapotranspiration (mm) using the Thornthwaite equation}

\item{EvpJun}{Mean Potential Evapotranspiration (mm) in June. Default uses values from Temp (can be created by Set_Temp) to calculate Potential Evapotranspiration (mm) using the Thornthwaite equation}

\item{EvpJul}{Mean Potential Evapotranspiration (mm) in July. Default uses values from Temp (can be created by Set_Temp) to calculate Potential Evapotranspiration (mm) using the Thornthwaite equation}

\item{EvpAug}{Mean Potential Evapotranspiration (mm) in August. Default uses values from Temp (can be created by Set_Temp) to calculate Potential Evapotranspiration (mm) using the Thornthwaite equation}

\item{EvpSep}{Mean Potential Evapotranspiration (mm) in September. Default uses values from Temp (can be created by Set_Temp) to calculate Potential Evapotranspiration (mm) using the Thornthwaite equation}

\item{EvpOct}{Mean Potential Evapotranspiration (mm) in October. Default uses values from Temp (can be created by Set_Temp) to calculate Potential Evapotranspiration (mm) using the Thornthwaite equation}

\item{EvpNov}{Mean Potential Evapotranspiration (mm) in November. Default uses values from Temp (can be created by Set_Temp) to calculate Potential Evapotranspiration (mm) using the Thornthwaite equation}

\item{EvpDec}{Mean Potential Evapotranspiration (mm) in December. Default uses values from Temp (can be created by Set_Temp) to calculate Potential Evapotranspiration (mm) using the Thornthwaite equation}
}
\value{
A dataframe with monthly mean evapotranspiration (mm).
}
\description{
The Set_Evp function create a dataframe with two columns and 12 rows. The first column indicates the Month (1-12). The second column indicates the monthly mean evapotranspirations (mm), either filled by the user or autofilled by default values calculated using the Thornthwaite equation using Temperature data from Set_Temp.
}
