% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Precip.R
\name{Set_Precip}
\alias{Set_Precip}
\title{Set Monthly Precipitation Vector}
\usage{
Set_Precip(
  PJan = 60.59,
  PFeb = 49.02,
  PMar = 59.08,
  PApr = 58.46,
  PMay = 56.81,
  PJun = 32.81,
  PJul = 18.48,
  PAug = 23.88,
  PSep = 45.09,
  POct = 79.58,
  PNov = 73.53,
  PDec = 70.59
)
}
\value{
A mvector with monthly precipitation values
}
\description{
The Set_Precip function creates a column with 12 rows and one column. The first
column indicates the Month (1-12). The second column indicates the mean Precipitation,
either filled by the user or autofilled by default values for Spain 1990-2020.
}
