% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Temp.R
\name{Set_Temp}
\alias{Set_Temp}
\title{Set Monthly Temperature Vector}
\usage{
Set_Temp(
  TJan = 6.26,
  TFeb = 7.17,
  TMar = 9.79,
  TApr = 11.82,
  TMay = 15.54,
  TJun = 19.95,
  TJul = 23.06,
  TAug = 23.09,
  TSep = 19.46,
  TOct = 14.83,
  TNov = 9.74,
  TDec = 6.95
)
}
\value{
A vector with monthly temperature values
}
\description{
The Set_Temp function create a column with 12 rows and one column. The first
column indicates the Month (1-12). The second column indicates the mean Temperature,
either filled by the user or autofilled by default values for Spain 1990-2020.
}
